+++ src/main/scala/draco/CLI.scala
@@ -10,15 +10,25 @@ object CLI extends DracoType {
   lazy val dracoType: Type[CLI[_]] = Type[CLI[_]] (typeDefinition)
   lazy val domainType: Domain[Draco] = Domain[Draco] (typeDefinition)
 
-  def load(path: String): TypeDefinition = {
+  def load(_path: => String): TypeDefinition = {
+    lazy val path: String = _path
     lazy val source: scala.io.BufferedSource = scala.io.Source.fromFile(path)
     lazy val text: String = try source.mkString finally source.close()
     io.circe.parser.parse(text).flatMap(_.as[TypeDefinition]).getOrElse(TypeDefinition.Null)
   }
   def version: Unit = println("Draco 2.0.0-alpha.1")
-  def inspect(path: String): Unit = println(TypeDefinition.encoder(load(path)).spaces2)
-  def generate(path: String): Unit = println(DracoGenerator.generate(load(path)))
-  def drake(path: String): Unit = println(Drake.emit(load(path)))
+  def inspect(_path: => String): Unit = {
+    lazy val path: String = _path
+    println(TypeDefinition.encoder(load(path)).spaces2)
+  }
+  def generate(_path: => String): Unit = {
+    lazy val path: String = _path
+    println(DracoGenerator.generate(load(path)))
+  }
+  def drake(_path: => String): Unit = {
+    lazy val path: String = _path
+    println(Drake.emit(load(path)))
+  }
   lazy val commands: Map[String, Seq[String] => Unit] = Map("version" -> (_ => version), "inspect" -> (as => inspect(as.head)), "generate" -> (as => generate(as.head)), "drake" -> (as => drake(as.head)))
   def main(args: Array[String]): Unit = commands.getOrElse(args.headOption.getOrElse("version"), (_: Seq[String]) => version)(args.drop(1).toSeq)
 }